8.24

+++ app/src/main/java/com/example/yangsong/piaoai/fragment/DayFragment.java
@@ -259,7 +259,7 @@ public class DayFragment extends BaseFragment implements OnChartValueSelectedLis
             AppUtil.TVOC(getActivity(), dayMsgTv, (double) e.getY());
         } else if (indext == 3) {
             //甲醛
-            AppUtil.jiaquan(getActivity(), dayMsgTv, (int) e.getY());
+            AppUtil.jiaquan(getActivity(), dayMsgTv,  e.getY());
         } else if (indext == 4) {
             //温度
             AppUtil.wendu(getActivity(), dayMsgTv, (int) e.getY());
@@ -285,7 +285,7 @@ public class DayFragment extends BaseFragment implements OnChartValueSelectedLis
             if (i >= (mList.size())) {
                 values1.add(new Entry(i, 0));
             } else
-                values1.add(new Entry(i, Integer.parseInt(mList.get(i))));
+                values1.add(new Entry(i, Float.parseFloat(mList.get(i))));
         }
