+++ app/src/main/java/com/example/yangsong/piaoai/fragment/MonthFragment.java
@@ -259,7 +259,7 @@ public class MonthFragment extends BaseFragment implements OnChartValueSelectedL
                 values1.add(new Entry(i, 0));
             } else {
 
-                values1.add(new Entry(i, Integer.parseInt(mList.get(i))));
+                values1.add(new Entry(i, Float.parseFloat(mList.get(i))));
             }
 
         }
@@ -303,7 +303,7 @@ public class MonthFragment extends BaseFragment implements OnChartValueSelectedL
             AppUtil.TVOC(getActivity(), dayMsgTv, (int) e.getY());
         } else if (indext == 3) {
             //甲醛
-            AppUtil.jiaquan(getActivity(), dayMsgTv, (int) e.getY());
+            AppUtil.jiaquan(getActivity(), dayMsgTv, e.getY());
         } else if (indext == 4) {
             //温度
             AppUtil.wendu(getActivity(), dayMsgTv, (int) e.getY());
