+++ app/src/main/java/com/example/yangsong/piaoai/fragment/PMDayFragment.java
@@ -195,7 +195,7 @@ public class PMDayFragment extends BaseFragment implements OnChartValueSelectedL
             if (i >= (mList.size())) {
                 values1.add(new Entry(i, 0));
             } else
-                values1.add(new Entry(i, Integer.parseInt(mList.get(i))));
+                values1.add(new Entry(i, Float.parseFloat(mList.get(i))));
         }
 
         LineDataSet set1;
