+++ app/src/main/java/com/example/yangsong/piaoai/fragment/PMWeekFragment.java
@@ -196,7 +196,7 @@ public class PMWeekFragment extends BaseFragment implements OnChartValueSelected
             if (i >= (mList.size())) {
                 values1.add(new Entry(i, 0));
             } else
-                values1.add(new Entry(i, Integer.parseInt(mList.get(i))));
+                values1.add(new Entry(i, Float.parseFloat(mList.get(i))));
         }
 
         LineDataSet set1;
