+++ app/src/main/java/com/example/yangsong/piaoai/fragment/WeekFragment.java
@@ -260,7 +260,7 @@ public class WeekFragment extends BaseFragment implements OnChartValueSelectedLi
             if (i >= (mList.size())) {
                 values1.add(new Entry(i, 0));
             } else
-                values1.add(new Entry(i, Integer.parseInt(mList.get(i))));
+                values1.add(new Entry(i, Float.parseFloat(mList.get(i))));
         }
         LineDataSet set1;
         if (mChart.getData() != null &&
@@ -301,7 +301,7 @@ public class WeekFragment extends BaseFragment implements OnChartValueSelectedLi
             AppUtil.TVOC(getActivity(), dayMsgTv, (int) e.getY());
         } else if (indext == 3) {
             //甲醛
-            AppUtil.jiaquan(getActivity(), dayMsgTv, (int) e.getY());
+            AppUtil.jiaquan(getActivity(), dayMsgTv, e.getY());
         } else if (indext == 4) {
             //温度
             AppUtil.wendu(getActivity(), dayMsgTv, (int) e.getY());
@@ -416,7 +416,7 @@ public class WeekFragment extends BaseFragment implements OnChartValueSelectedLi
             default:
                 break;
         }
-        // initYLabel(position);
+        //initYLabel(position);
     }
 
 }
